Refactoring and new Spec for heap optimization
