+++ src/test/scala/testing/optimiseheap.scala
@@ -7,10 +7,12 @@ import _root_.scala.Predef._
 //   optimiseHeapOptions = true,
 //   optimiseHeapBooleans = true,
 //   optimiseHeapStrings = true,
+//   optimisePrimitives = true
 // )
-// class Foo(a: Option[Boolean], b: Option[Boolean], s: Option[String])
+// class Foo(a: Option[Boolean], b: Option[Boolean], s: Option[String], i: Int)
 final class Foo private (
   private[this] var _bitmask: Long,
+  private[this] var _bytes: Array[Byte],
   private[this] var _s: Array[Byte]
 ) extends scala.Serializable {
 
@@ -23,21 +25,31 @@ final class Foo private (
   def s: Option[String] =
     if (_s == null) None
     else Some(new String(_s)) // optimiseHeapStrings
+  def i: Int = {
+    var _i: Int = 0
+    _i |= (_bytes(3) << 24)
+    _i |= (_bytes(2) << 16)
+    _i |= (_bytes(1) << 8)
+    _i |= (_bytes(0) << 0)
+    _i
+  } // optimisePrimitives
 
   def copy(
     a: Option[Boolean] = a,
     b: Option[Boolean] = b,
-    s: Option[String] = s
-  ): Foo = Foo(a, b, s)
+    s: Option[String] = s,
+    i: Int = i
+  ): Foo = Foo(a, b, s, i)
 
-  override def toString(): String = s"Foo($a,$b,$s)"
+  override def toString(): String = s"Foo($a,$b,$s,$i)"
   // I'm not going to lie, hashCode/equals is slower... this is about trading CPU for heap
   // (note that memoiseHashCode could be done before construction to avoid unpacking)
   override def hashCode(): Int =
-    a.hashCode + 13 * (b.hashCode + 13 * s.hashCode)
+    a.hashCode + 13 * (b.hashCode + 13 * (s.hashCode + 13 * i.hashCode()))
   override def equals(o: Any): Boolean = o match {
-    case that: Foo => (this eq that) || (a == that.a && s == that.s)
-    case _         => false
+    case that: Foo =>
+      (this eq that) || (a == that.a && b == that.b && s == that.s && i == that.i)
+    case _ => false
   }
 
   @throws[java.io.IOException]
@@ -46,6 +58,7 @@ final class Foo private (
     out.writeObject(a: Serializable)
     out.writeObject(b: Serializable)
     out.writeObject(s: Serializable)
+    out.writeInt(i)
   }
   @throws[java.io.IOException]
   @throws[java.lang.ClassNotFoundException]
@@ -53,17 +66,19 @@ final class Foo private (
     val a      = in.readObject().asInstanceOf[Option[Boolean]]
     val b      = in.readObject().asInstanceOf[Option[Boolean]]
     val s      = in.readObject().asInstanceOf[Option[String]]
-    val packed = Foo.pack(a, b, s)
+    val i      = in.readInt()
+    val packed = Foo.pack(a, b, s, i)
     _bitmask = packed._1
-    _s = packed._2
+    _bytes = packed._2
+    _s = packed._3
   }
   @throws[java.io.ObjectStreamException]
-  private[this] def readResolve(): Any = Foo(a, b, s)
+  private[this] def readResolve(): Any = Foo(a, b, s, i)
 
 }
 
 final object Foo
-    extends ((Option[Boolean], Option[Boolean], Option[String]) => Foo)
+    extends ((Option[Boolean], Option[Boolean], Option[String], Int) => Foo)
     with scala.Serializable {
   override def toString = "Foo"
 
@@ -78,7 +93,8 @@ final object Foo
   // unpack is available on the instance, i.e. field access
   private def pack(a: Option[Boolean],
                    b: Option[Boolean],
-                   s: Option[String]): (Long, Array[Byte]) = {
+                   s: Option[String],
+                   i: Int): (Long, Array[Byte], Array[Byte]) = {
     var _bitmask: Long = 0L
 
     if (a == None) _bitmask |= (1 << 0)
@@ -89,18 +105,27 @@ final object Foo
 
     val _s = if (s == None) null else s.get.getBytes
 
-    (_bitmask, _s)
+    val _bytes: Array[Byte] = Array.fill(4)(0)
+    _bytes(0) = i.toByte
+    _bytes(1) = (i >> 8).toByte
+    _bytes(2) = (i >> 16).toByte
+    _bytes(3) = (i >> 24).toByte
+
+    (_bitmask, _bytes, _s)
   }
 
-  def apply(a: Option[Boolean], b: Option[Boolean], s: Option[String]): Foo = {
-    val packed = pack(a, b, s)
-    val foo    = new Foo(packed._1, packed._2)
+  def apply(a: Option[Boolean],
+            b: Option[Boolean],
+            s: Option[String],
+            i: Int): Foo = {
+    val packed = pack(a, b, s, i)
+    val foo    = new Foo(packed._1, packed._2, packed._3)
     foo.synchronized(foo) // safe publish
   }
   def unapply(
     f: Foo
-  ): Option[(Option[Boolean], Option[Boolean], Option[String])] =
-    Some((f.a, f.b, f.s))
+  ): Option[(Option[Boolean], Option[Boolean], Option[String], Int)] =
+    Some((f.a, f.b, f.s, f.i))
 
   import shapeless.{ ::, Generic, HNil, LabelledGeneric, Typeable }
   import shapeless.labelled.{ field, FieldType }
@@ -108,13 +133,14 @@ final object Foo
   val a_tpe = 'a.narrow
   val b_tpe = 'b.narrow
   val s_tpe = 's.narrow
+  val i_tpe = 'i.narrow
   implicit val TypeableFoo: Typeable[Foo] = new Typeable[Foo] {
     override def cast(t: Any): Option[Foo] = t match {
       case f: Foo => Some(f) // no type params, so trivial
       case _      => None
     }
     override def describe: String =
-      s"Foo[Option[Boolean],Option[Boolean],Option[String]]"
+      s"Foo[Option[Boolean],Option[Boolean],Option[String],Int]"
   }
 
   implicit val LabelledGenericFoo
@@ -122,29 +148,32 @@ final object Foo
                           FieldType[a_tpe.type, Option[Boolean]] ::
                             FieldType[b_tpe.type, Option[Boolean]] ::
                             FieldType[s_tpe.type, Option[String]] ::
+                            FieldType[i_tpe.type, Int] ::
                             HNil] =
     new LabelledGeneric[Foo] {
       override type Repr =
         FieldType[a_tpe.type, Option[Boolean]] ::
           FieldType[b_tpe.type, Option[Boolean]] ::
           FieldType[s_tpe.type, Option[String]] ::
+          FieldType[i_tpe.type, Int] ::
           HNil
       override def to(f: Foo): Repr =
         field[a_tpe.type](f.a) ::
           field[b_tpe.type](f.b) ::
           field[s_tpe.type](f.s) ::
+          field[i_tpe.type](f.i) ::
           HNil
       override def from(r: Repr): Foo = GenericFoo.from(r)
     }
 
-  implicit val GenericFoo
-    : Generic.Aux[Foo, Option[Boolean] :: Option[Boolean] :: Option[String] :: HNil] =
+  implicit val GenericFoo: Generic.Aux[Foo, Option[Boolean] :: Option[Boolean]
+    :: Option[String] :: Int :: HNil] =
     new Generic[Foo] {
       override type Repr =
-        Option[Boolean] :: Option[Boolean] :: Option[String] :: HNil
+        Option[Boolean] :: Option[Boolean] :: Option[String] :: Int :: HNil
       override def to(f: Foo): Repr = LabelledGenericFoo.to(f)
       override def from(r: Repr): Foo = r match {
-        case a :: b :: s :: HNil => Foo(a, b, s)
+        case a :: b :: s :: i :: HNil => Foo(a, b, s, i)
       }
     }
